Refactor Auth / Layout
