+++ src/app/layout.tsx
@@ -22,31 +22,23 @@
  */
 'use client';
 
-import { ReactNode, useState } from 'react';
+import { ReactNode } from 'react';
 import { QueryClient, QueryClientProvider } from 'react-query';
 import { Work_Sans } from 'next/font/google';
-import { usePathname } from 'next/navigation';
 
 import { AuthProvider } from '@/global/utils/auth';
 import Header from './components/Header';
 import ThemeProvider from './components/ThemeProvider';
-import { getToken } from '@/global/utils/auth';
 
 const workSans = Work_Sans({ subsets: ['latin'] });
 
 export default function RootLayout({ children }: { children: ReactNode }) {
-	const path = usePathname();
-	const storedToken = getToken();
-	const [egoJwt, setEgoJwt] = useState(storedToken || '');
-	const initLoginState = path === '/logging-in' && !(storedToken || egoJwt) ? true : false;
-	const [loggingIn, setLoggingIn] = useState(initLoginState);
-
 	return (
 		<html lang="en">
 			<body className={workSans.className}>
 				<ThemeProvider>
 					<QueryClientProvider client={new QueryClient()}>
-						<AuthProvider authData={{ egoJwt, setEgoJwt, loggingIn, setLoggingIn }}>
+						<AuthProvider>
 							<Header />
 							{children}
 						</AuthProvider>
